update: added "accepted" attribute to articles
We've decided that in order for moderators to revisit their decision to accept/reject article suggestions, there should be a field in the database for the "accepted" status of each one. This change adds a new "accepted" attribute to each article suggestion when it is added to the database. It is "false" by default and cannot be changed by the user, only the moderator